cse4310 homework #1 (#5)
* extract data for homework 1

* move data folder from homework folder

* fulfill part 1

* remove intermediate images, update comments

* fix problem with bbox on frames 193-195

* extract person finding part of find_bounding_box

This is because task 3 also requires this functionality, but shouldn't
display any images. So now find_bounding_box just calls find_person and
then draws the frame around the person and displays it.

* fulfill task 3

* made find_person a lot faster

regionprops can find the size of each region in the labeled image

* improve calculating bounding box algorithm

* implement person_present function

* add readme

* move README to correctly location

* remove unnecessary function in find_person.m

* use all(diff(:) == 0) instead of sum(diff(:)) == 0

* switch to using minmax instead of min then max

* increase threshold and add dilation

The increased threshold reduces the size of the bounding box later in
the the sequence, while the dilation allows for the person to reconnect
with his legs on frame 95.